Add FileBasedStorage provider for JVM and improve Android provider
